+++ tools/testproducer/testproducer.go
@@ -6,14 +6,15 @@ import (
 	"encoding/base64"
 	"flag"
 	"fmt"
-	"io/ioutil"
-	"net"
-	"net/http"
-	"strings"
+	"strconv"
 	"sync"
 	"time"
 
+	pb "github.com/mailgun/kafka-pixy/gen/golang"
 	"github.com/mailgun/kafka-pixy/prettyfmt"
+	"github.com/pkg/errors"
+	"golang.org/x/net/context"
+	"google.golang.org/grpc"
 )
 
 const (
@@ -21,7 +22,7 @@ const (
 )
 
 var (
-	pixyAddr string
+	grpcAddr string
 	topic    string
 	isSync   bool
 	threads  int
@@ -30,11 +31,11 @@ var (
 )
 
 func init() {
-	flag.StringVar(&pixyAddr, "addr", "localhost:19092", "either unix domain socker or TCP address")
-	flag.StringVar(&topic, "topic", "test", "the name of the topic")
+	flag.StringVar(&grpcAddr, "addr", "localhost:19091", "gRPC server address")
+	flag.StringVar(&topic, "topic", "test", "name of the topic")
 	flag.BoolVar(&isSync, "sync", false, "should production be synchronous")
 	flag.IntVar(&threads, "threads", 1, "number of concurrent producer threads")
-	flag.IntVar(&count, "count", 10000, "number of messages to produce")
+	flag.IntVar(&count, "count", 10000, "number of messages to produce by all threads")
 	flag.IntVar(&size, "size", 1000, "message size in bytes")
 	flag.Parse()
 }
@@ -43,22 +44,18 @@ func main() {
 	msg := genmessage(size)
 	progressCh := make(chan int)
 
-	useUnixDomainSocket := false
-	var baseURL string
-	if strings.HasPrefix(pixyAddr, "/") {
-		fmt.Printf("Using UDS client for %s\n", pixyAddr)
-		baseURL = "http://_"
-		useUnixDomainSocket = true
-	} else {
-		fmt.Printf("Using net client for %s\n", pixyAddr)
-		baseURL = fmt.Sprintf("http://%s", pixyAddr)
+	cltConn, err := grpc.Dial(grpcAddr, grpc.WithInsecure())
+	if err != nil {
+		panic(errors.Wrap(err, "failed to dial gRPC server"))
 	}
+	clt := pb.NewKafkaPixyClient(cltConn)
 
 	go func() {
 		var wg sync.WaitGroup
 		chunkSize := count / threads
-		for i := 0; i < threads; i++ {
-			messageIndexBegin := chunkSize * i
+		for tid := 0; tid < threads; tid++ {
+			tid := tid
+			messageIndexBegin := chunkSize * tid
 			messageIndexEnd := messageIndexBegin + chunkSize
 			if count-messageIndexEnd < chunkSize {
 				messageIndexEnd = count
@@ -67,39 +64,24 @@ func main() {
 			go func() {
 				defer wg.Done()
 
-				var clt http.Client
-				if useUnixDomainSocket {
-					dial := func(proto, addr string) (net.Conn, error) {
-						return net.Dial("unix", pixyAddr)
-					}
-					clt.Transport = &http.Transport{Dial: dial}
-				}
-
 				recentProgress := 0
 				checkpoint := time.Now()
 				for i := messageIndexBegin; i < messageIndexEnd; i++ {
-					var URL string
-					if isSync {
-						URL = fmt.Sprintf("%s/topics/%s/messages?sync&key=%d", baseURL, topic, i)
-					} else {
-						URL = fmt.Sprintf("%s/topics/%s/messages?key=%d", baseURL, topic, i)
+					req := pb.ProdReq{
+						Topic:     topic,
+						KeyValue:  []byte(strconv.Itoa(i)),
+						Message:   msg,
+						AsyncMode: !isSync,
 					}
-					res, err := clt.Post(URL, "text/plain", bytes.NewReader(msg))
+					_, err := clt.Produce(context.Background(), &req)
 					if err != nil {
-						panic(fmt.Errorf("failed to POST %s, err=(%s)", URL, err))
+						panic(errors.Wrapf(err, "failed to produce: thread=%d, no=%d", tid, i-messageIndexBegin))
 					}
-					if res.StatusCode != http.StatusOK {
-						body, err := ioutil.ReadAll(res.Body)
-						if err != nil {
-							panic(err)
-						}
-						panic(fmt.Sprintf("%d %s", res.StatusCode, body))
-					}
-					res.Body.Close()
 					recentProgress += 1
 					if time.Now().Sub(checkpoint) > reportingPeriod {
 						progressCh <- recentProgress
 						recentProgress = 0
+						checkpoint = time.Now()
 					}
 				}
 				progressCh <- recentProgress
@@ -130,12 +112,12 @@ func main() {
 func genmessage(size int) []byte {
 	raw := make([]byte, size)
 	if _, err := rand.Read(raw); err != nil {
-		panic(fmt.Sprintf("failed to generate message: err=(%s)", err))
+		panic(errors.Wrap(err, "failed to generate message"))
 	}
 	var buf bytes.Buffer
 	encoder := base64.NewEncoder(base64.StdEncoding, &buf)
 	if n, err := encoder.Write(raw); err != nil || n != len(raw) {
-		panic(fmt.Sprintf("failed to encode message: err=(%s)", err))
+		panic(errors.Wrap(err, "failed to encode message"))
 	}
 	encoded := buf.Bytes()
 	if len(encoded) > size {